clean up and can run from command line command without editing file
